Convert from using `float` to `double`
Resolves #150

+++ src/InOneWeekend/hittable.h
@@ -18,7 +18,7 @@ class material;
 
 struct hit_record
 {
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
     material *mat_ptr;
@@ -26,7 +26,7 @@ struct hit_record
 
 class hittable  {
     public:
-        virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+        virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
 };
