+++ src/InOneWeekend/hittable_list.h
@@ -18,12 +18,12 @@ class hittable_list: public hittable  {
     public:
         hittable_list() {}
         hittable_list(hittable **l, int n) { list = l; list_size = n; }
-        virtual bool hit(const ray& r, float tmin, float tmax, hit_record& rec) const;
+        virtual bool hit(const ray& r, double tmin, double tmax, hit_record& rec) const;
         hittable **list;
         int list_size;
 };
 
-bool hittable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
+bool hittable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const {
     hit_record temp_rec;
     bool hit_anything = false;
     double closest_so_far = t_max;
