+++ src/TheNextWeek/ray.h
@@ -18,15 +18,15 @@ class ray
 {
     public:
         ray() {}
-        ray(const vec3& a, const vec3& b, float ti = 0.0) { A = a; B = b; _time = ti;}  
-        vec3 origin() const       { return A; }
-        vec3 direction() const    { return B; }
-        float time() const    { return _time; }
-        vec3 point_at_parameter(float t) const { return A + t*B; }
+        ray(const vec3& a, const vec3& b, double ti = 0.0) { A = a; B = b; _time = ti;}  
+        vec3 origin() const    { return A; }
+        vec3 direction() const { return B; }
+        double time() const    { return _time; }
+        vec3 point_at_parameter(double t) const { return A + t*B; }
 
         vec3 A;
         vec3 B;
-        float _time;
+        double _time;
 };
